+++ src/commands/index.ts
@@ -2,6 +2,7 @@ import { Command } from "commander";
 import { registerDefaultCommand } from "./default.js";
 import { registerAnalyzeCommand } from "./analyze.js";
 import { registerReportCommand } from "./report.js";
+import { registerExportCommand } from "./export.js";
 import { registerCompressCommand } from "./compress.js";
 import { registerNewChatCommand } from "./new-chat.js";
 import { registerInitCommand } from "./init.js";
@@ -17,6 +18,7 @@ export function registerCommands(program: Command, options: RegisterCommandsOpti
   registerInitCommand(program, { packageRoot: options.packageRoot });
   registerAnalyzeCommand(program);
   registerReportCommand(program);
+  registerExportCommand(program);
   registerCompressCommand(program);
   registerNewChatCommand(program, { packageVersion: options.packageVersion });
   program.addCommand(createHookCommand(), { hidden: true });